Change gasnet_node_t - GASNet implementors please read carefully
Change the default definition of gasnet_node_t to be uint32_t instead of uint8_t, to allow scaling beyond 255 nodes. Conduits are still free to override this definition in gasnet_core_fwd.h, although currently all of them are using the default definition except for gm-conduit which defines it to be uint16_t (and should probably be left that way because it currently makes some assumptions about this size, and even some non-64-bit compliant assumptions in gm-conduit/gasnet_core_conf.c:158, which Christian should fix)

I've tested this change throughly on the main-line conduits, and as far as I can tell the only significant impact is that lapi-conduit's gasnetc_msg_t (on-the-wire AM header) is now 4-bytes longer (this could be compressed back to the original size by imposing a 255-node max, but given the current overheads with AM on lapi-conduit it's unlikely to even make any noticeable difference - other compression could be done to this header, for example gasnetc_flag_t is currently wasting lots of bits in the header.. ).

This fixes bug 215 for AMMPI and GASNet

Clean up some sloppy use of sched_yield(), which is not available on all systems.

+++ gasnet.h
@@ -1,6 +1,6 @@
 /*  $Archive:: /Ti/GASNet/gasnet.h                                        $
- *     $Date: 2003/04/17 11:48:00 $
- * $Revision: 1.12 $
+ *     $Date: 2003/05/22 09:21:20 $
+ * $Revision: 1.13 $
  * Description: GASNet Header
  * Copyright 2002, Dan Bonachea <bonachea@cs.berkeley.edu>
  * Terms of use are as specified in license.txt
@@ -154,7 +154,7 @@
 #ifndef _GASNET_NODE_T
 #define _GASNET_NODE_T
   /*  unsigned integer type representing a unique 0-based node index */
-  typedef uint8_t gasnet_node_t;
+  typedef uint32_t gasnet_node_t;
 #endif
 
 #ifndef _GASNET_HANDLER_T
